Update Dual Camera Control.vi
Must be > 0 "Wait on Notifications" waiting before sending the preview or capture command. Prevent the user from changing the number of frames or exposure time but not having triggered a value change event and immediately pressing the preview or capture button (this leaves the button TRUE but the event is notification is ignored).